+++ NEWS.md
@@ -3,6 +3,7 @@ apexcharter 0.4.1
 
 * Updated ApexCharts.js to 3.40.0 (new chart type: dumbbell chart).
 * `apex()` : added support for boxplot.
+* New function `ax_forecast_data_points()` to mark points as forecasted values.